Iris Dataset
The training, test and validation data are saved in this .csv file. 

There are 150 samples. Each sample has 4 inputs (sepal width, sepal length, petal width, petal length) and 1 output (species of flower; setosa, versicolor or virginica).

The data is ordered so that the first 50 samples are for the setosa flower, the 50 after that are for the versicolor flower and the 50 after that are for the virgnica flower.