Label encoding
The purpose of this code is to:

•	Read the csv file and put all the contents into a cell array called textdata
•	Put the first 4 columns (the 4 integer inputs) from textdata into a numeric matrix called data
•	Put the 5th column (the string outputs) from textdata into a vector called target
•	Do label encoding – assign each string value from target a numerical value of -1,0 or 1
•	Store these new numerical values into a matrix called tr
•	Append the matrix tr to the matrix data – we now have a single matrix called data  which contains the 4 integer inputs and 1 integer output (which used to be a string output)

When representing different classes, such as setosa, versicolor, and virginica, it is best practice to assign integers to each class instead of keeping them as string values. E.g. setosa = -1, versicolor = 0, virginica = 1. It does not matter which is which as long as it's consistent and a key/record is kept. This process is called "label encoding".