LDA Model
The purpose of this code is to perform linear discriminant analysis (LDA) on the Fisher's Iris dataset. It reads the dataset from a CSV file, trains an LDA classifier using the first four numeric columns as features and the last categorical column as the target variable ('Species'). Then, it makes predictions on the same dataset and evaluates the classification performance using a confusion matrix. The overall goal is to demonstrate and assess the accuracy of the LDA classifier in predicting the species of iris flowers based on their features.