LM Model
There are 2 purposes of this code:
1. An alternative way to train the model. This code trains the feedforward neural network using the Levenberg-Marquardt (LM) backpropagation algorithm.
2. Evaluate the performance of the model with varying topologies (number of hidden layers).

In this code, the iris dataset is read and loaded using the Neural Network Toolbox. Then a nested for loop is executed to explore the performance of the neural network with different hidden layer sizes/number of neurons and the code is run for each configuration/combination of hidden layers multiple times to get an average to account for variability. Lower values in the matrix perfMat indicate a lower error and therefore a higher-performing model.